requirements.txt
flask
pyqt6
supabase
pandas
python-dotenv
flask-oauthlib
openai
authlib
python-dotenv